+++ src/hooks.ts
@@ -1,9 +1,8 @@
-import * as cookie from 'cookie'
 import type { Handle, RequestEvent } from '@sveltejs/kit'
 import { query } from './routes/_db'
 
 // Attach authorization to each server request (role may have changed)
-async function attachUserToRequest(sessionId: string, event: RequestEvent) {
+async function attachUserToRequestEvent(sessionId: string, event: RequestEvent) {
   const sql = `
     SELECT * FROM get_session($1);`
   const { rows } = await query(sql, [sessionId])
@@ -12,24 +11,20 @@ async function attachUserToRequest(sessionId: string, event: RequestEvent) {
   }
 }
 
-function deleteCookieIfNoUser(event: RequestEvent, response: Response) {
-  if (!event.locals.user) {
-    response.headers.set('Set-Cookie', `session=; Path=/; HttpOnly; SameSite=Lax; Expires=${new Date().toUTCString()}`)
-  }
-}
-
 // Invoked for each endpoint called and initially for SSR router
 export const handle: Handle = async ({ event, resolve }) => {
+  const { cookies } = event
+  const sessionId = cookies.get('session')
 
   // before endpoint or page is called
-  const cookies = cookie.parse(event.request.headers.get('Cookie') || '')
-  if (cookies.session) {
-    await attachUserToRequest(cookies.session, event)
+  if (sessionId) {
+    await attachUserToRequestEvent(sessionId, event)
   }
 
   const response = await resolve(event)
 
   // after endpoint or page is called
-  deleteCookieIfNoUser(event, response)
+  if (!event.locals.user) cookies.delete('session')
+
   return response
 }