+++ src/routes/admin/+page.server.ts
@@ -4,8 +4,9 @@ import type { PageServerLoad } from './$types'
 export const load: PageServerLoad = async ({locals})=> {
 	const { user } = locals
 	const authorized = ['admin']
-	if (user && !authorized.includes(user.role)) {
-		throw redirect(302, '/login?referrer=/admin');
+	console.log('admin/+page.server.ts', user)
+	if (!user || !authorized.includes(user.role)) {
+		throw redirect(302, '/login?referrer=/admin')
 	}
 
 	return {