+++ src/routes/profile/+page.server.ts
@@ -5,7 +5,7 @@ export const load: PageServerLoad = async ({ locals }) => {
   const { user } = locals // populated by /src/hooks.ts
 
   const authorized = ['admin', 'teacher', 'student'] // must be logged-in
-  if (user && !authorized.includes(user.role)) {
+  if (!user || !authorized.includes(user.role)) {
     throw redirect(302, '/login?referrer=/profile')
   }