+++ src/routes/teachers/+page.server.ts
@@ -2,8 +2,9 @@ import { redirect } from '@sveltejs/kit'
 import type { PageServerLoad } from './$types'
 
 export const load: PageServerLoad = async ({locals}) => {
+	const { user } = locals
 	const authorized = ['admin', 'teacher']
-	if (!locals.user || !authorized.includes(locals.user.role)) {
+	if (!user || !authorized.includes(user.role)) {
 		throw redirect(302, '/login?referrer=/teachers')
 	}